Move funm_vector_product from matfun.py to lanczos.py
